% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Correlation Matrix Function}
\usage{
corr_matrix(x, y)
}
\description{
This function allows you to produce a correlation matrix for numerical variables in a particular year.
}
\examples{
corr_matrix()
}
\keyword{NBA}
