% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_points.R
\name{max_points}
\alias{max_points}
\title{A Points Function}
\usage{
max_points(x, y)
}
\description{
This function allows you to reveal the NBA player with most points in a given year.
}
\examples{
max_points()
}
\keyword{NBA}
