% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(x, y)
}
\description{
This function allows you to reveal the oldest NBA player in a particular year.
}
\examples{
oldest_player()

}
\keyword{NBA}
